Updated - 18th march 3.01pm - from ZL lap. (Added participated events view )

+++ src/main/java/com/project/sportsManagement/entity/Student.java
@@ -100,7 +100,7 @@ public class Student implements UserDetails {
         this.participation = new HashSet<>();
     }
 
-    public long getStudentId() {
+    public int getStudentId() {
         return studentId;
     }
